+++ main.tf
@@ -1,3 +1,7 @@
+locals {
+  effective_tags = length(var.es_tags) > 0 ? var.es_tags : var.tags
+}
+
 module "tamr-es-cluster" {
   source                          = "./modules/aws-es"
   domain_name                     = var.domain_name
@@ -11,7 +15,7 @@ module "tamr-es-cluster" {
   ebs_iops                        = var.ebs_iops
   ebs_volume_size                 = var.ebs_volume_size
   ebs_volume_type                 = var.ebs_volume_type
-  additional_tags                 = var.additional_tags
+  tags                            = var.effective_tags
   aws_region                      = var.aws_region
   create_new_service_role         = var.create_new_service_role
   kms_key_id                      = var.kms_key_id